% Auto-generated: do not edit by hand
\name{page}

\alias{page}

\title{Page component}

\description{

}

\usage{
page(children=NULL, id=NULL, size=NULL, style=NULL)
}

\arguments{
\item{children}{Unnamed list. }

\item{id}{Character. }

\item{size}{Character. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

