% Auto-generated: do not edit by hand
\name{view}

\alias{view}

\title{View component}

\description{

}

\usage{
view(children=NULL, id=NULL, style=NULL)
}

\arguments{
\item{children}{Unnamed list. }

\item{id}{Character. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

